IssueID#18684 - Split map config from client config:
Split the existing client side configuration into two: a client config for
specifying interface components and a separate map config.

The map config is loaded via a service (map-config.service) which is currently
hardcoded: this will be changed to retrieve the data from the server in a later
commit.

The map.component now also gets the collection ID via the ActivatedRoute's
data.collectionId property, which is used to retrieve the correct map
configuration.