IssueID#18684 - Split map config from client config:
Split off the client config which controls the interface from the OpenLayers
map config, resulting in two services. The two services were not combined
because the map config requires a request to the server to get it's config
while the client config is "static" and client-side only.

The map config has been used to the test proxy server, although it's format is
still to be updated.