IssueID#18684 - Load map configuration from server:
Loads the map configuration from the server, converting roam's current map
config file into a new format to be used by the frontend.

This commit gets maps onto the interface, but does contain various parts which
will need to be updated, e.g. retrieving mapproxy urls from the global config.

Interfaces were created to match the current map config to make the above job
slightly easier, but may be removed when or if the backend config is updated.